+++ src/js/models/particle.js
