+++ src/js/models/shipsImgSource.js
