--- src/js/utils/playRandSound.js
@@ -1,6 +1,0 @@
-export default function playRandSound(arr) {
-  const audio = arr[Math.floor(Math.random() * arr.length)];
-  audio.pause();
-  audio.currentTime = 0;
-  audio.play();
-}
