--- src/js/utils/stopSound.js
@@ -1,5 +1,0 @@
-export default function stopSound(audio) {
-  const sound = audio;
-  sound.pause();
-  sound.loop = false;
-}
