+++ src/js/view/pageLayoutElements.js
@@ -1,5 +1,6 @@
 export default {
-  mainMenu: `<div class="main-menu">
+  mainMenu: `
+  <div class="main-menu">
     <ul class="main-menu__list">
       <li class="main-menu__item">New Game</li>
       <li class="main-menu__item">Continue</li>
@@ -9,22 +10,25 @@ export default {
       <li class="main-menu__item">Options</li>
     </ul>
   </div>`,
-  modal: `<div class="modal">
-  <div class="modal__inner">
-  <h2 class="modal__message">GAME OVER</h2>
-    <p class="modal__score">0</p>
-    <p class="modal__text">Points</p>
-  <div class="profile-menu__btn-wrapper">
-    <button class="profile-menu__btn-nav btn btn-main-menu">Main menu</button>
-    <button class="profile-menu__btn-nav btn btn-start-game">Start game</button>
-  </div>
-  </div>
-</div>`,
-  score: `<div class="score">
-  <span>Score: </span><span id="scoreEl">0</span>
+  modal: `
+  <div class="modal">
+    <div class="modal__inner">
+      <h2 class="modal__message">GAME OVER</h2>
+      <p class="modal__score">0</p>
+      <p class="modal__text">Points</p>
+      <div class="profile-menu__btn-wrapper">
+        <button class="profile-menu__btn-nav btn btn-main-menu">Main menu</button>
+        <button class="profile-menu__btn-nav btn btn-start-game">Start game</button>
+      </div>
+    </div>
   </div>`,
-  health: `<div class="health">
-  <span>Health: </span><span id="healthEl">3</span>
+  score: `
+  <div class="score">
+    <span>Score: </span><span id="scoreEl">0</span>
+  </div>`,
+  health: `
+  <div class="health">
+    <span>Health: </span><span id="healthEl">3</span>
   </div>`,
   profileMenu: `
   <div class="profile-menu wrapper">
@@ -37,14 +41,14 @@ export default {
           <div class="profile-menu__img"></div>
         </div>
         <div class="profile-menu__description">
-          <h3 class="profile-menu__ship-name">Falcone</h3>
+          <h3 class="profile-menu__ship-name"></h3>
           <p class="profile-menu__text">Speed:
           <span class="profile-menu__ship-speed"></span> Health:
           <span class="profile-menu__ship-health"></span> Laser:
           <span class="profile-menu__ship-laser"></span>
           </p>
         </div>
-      </div>
+        </div>
         <button class="profile-menu__btn-ship btn-next">></button>
       </div>
       <div class="profile-menu__form">
@@ -85,11 +89,15 @@ export default {
     <h2 class="top-page__title title">Top list</h2>
     <div class="container">
       <table class="top-page__table">
-        <tr>
-          <th class="top-page__table-header">Position</th>
-          <th class="top-page__table-header">Name</th>
-          <th class="top-page__table-header">Score</th>
-        </tr>
+        <thead>
+          <tr>
+            <th class="top-page__table-header">Position</th>
+            <th class="top-page__table-header">Name</th>
+            <th class="top-page__table-header">Score</th>
+          </tr>
+        </thead>
+        <tbody class="top-page__tbody">
+        </tbody>
       </table>
     </div>
     <div class="btn-wrapper">
@@ -105,6 +113,11 @@ export default {
         <p class="about__text">Vitali Burakou</p>
         <a class="about__link" href="mailto:vitali.burakou@gmail.com"><img class="about__icon" src="https://www.kindpng.com/picc/m/164-1649451_email-transparency-portable-network-graphics-computer-email-icons.png" width="20">vitali.burakou@gmail.com</a>
         <a class="about__link" href="#"><img class="about__icon" src="https://th.bing.com/th/id/OIP.H78e62IgOKHqLmIDbTN4igHaHa?w=161&h=180&c=7&o=5&pid=1.7" width="20">https://github.com/Vitali-Minsk</a>
+      </div>
+      <div class="about__info-container">
+        <p class="about__text">Dzmitry Palyka</p>
+        <a class="about__link" href="mailto:nixonstars@mail.ru"><img class="about__icon" src="https://www.kindpng.com/picc/m/164-1649451_email-transparency-portable-network-graphics-computer-email-icons.png" width="20">nixonstars@mail.ru</a>
+        <a class="about__link" href="#"><img class="about__icon" src="https://th.bing.com/th/id/OIP.H78e62IgOKHqLmIDbTN4igHaHa?w=161&h=180&c=7&o=5&pid=1.7" width="20">https://github.com/spanderok</a>
       </div>
       <div class="about__info-container">
         <p class="about__text">RSSchool</p>
