+++ src/styles/mainMenu.scss
@@ -1,6 +1,3 @@
-@import "roboto.css";
-@import "~normalize.css";
-
 .main-menu {
   width: 100%;
   height: 100%;
