+++ src/styles/modal.scss
@@ -1,9 +1,5 @@
-
 .modal {
   position: fixed;
-  // left: 50%;
-  // top: 50%;
-  // transform: translate(-50%, -50%);
   display: flex;
   justify-content: center;
   align-items: center;
@@ -39,7 +35,3 @@
   align-items: center;
   width: 450px;
 }
-
-// .modal_hidden {
-//   display: none;
-// }
