+++ src/styles/startPage.scss
