+++ src/styles/style.scss
@@ -27,7 +27,6 @@ html, body{
 }
 
 .container {
-  // width: 100;
   height: 100vh;
   display: flex;
   flex-direction: column;
@@ -65,4 +64,9 @@ html, body{
 
 .btn:hover {
   cursor: pointer;
+}
+
+.btn:active {
+  cursor: pointer;
+  color: rgb(156, 156, 156);
 }
